fix: cast adjust to reduce casting

+++ src/accumulators/large_accumulator.rs
@@ -7,26 +7,26 @@ use crate::{
 };
 
 pub(crate) struct LargeAccumulator {
-    pub(crate) m_chunk: [u64; XSUM_LCHUNKS as usize], // Chunks making up large accumulator
-    pub(crate) m_count: [i32; XSUM_LCHUNKS as usize], // Counts of # adds remaining for chunks, or -1 if not used yet or special
-    pub(crate) m_chunks_used: [u64; (XSUM_LCHUNKS / 64) as usize], // Bits indicate chunks in use
-    pub(crate) m_used_used: u64,                      // Bits indicate chunk_used entries not 0
-    pub(crate) m_sacc: SmallAccumulator,              // The small accumulator to condense into
+    pub(crate) m_chunk: [u64; XSUM_LCHUNKS], // Chunks making up large accumulator
+    pub(crate) m_count: [i32; XSUM_LCHUNKS], // Counts of # adds remaining for chunks, or -1 if not used yet or special
+    pub(crate) m_chunks_used: [u64; XSUM_LCHUNKS / 64], // Bits indicate chunks in use
+    pub(crate) m_used_used: u64,             // Bits indicate chunk_used entries not 0
+    pub(crate) m_sacc: SmallAccumulator,     // The small accumulator to condense into
 }
 impl LargeAccumulator {
     pub(crate) fn new() -> Self {
         //  : m_chunk(XSUM_LCHUNKS), m_count(XSUM_LCHUNKS, -1), m_chunksUsed(XSUM_LCHUNKS / 64, 0), m_usedUsed{0}, m_sacc{}
         Self {
-            m_chunk: [0; XSUM_LCHUNKS as usize],
-            m_count: [-1; XSUM_LCHUNKS as usize],
-            m_chunks_used: [0; (XSUM_LCHUNKS / 64) as usize],
+            m_chunk: [0; XSUM_LCHUNKS],
+            m_count: [-1; XSUM_LCHUNKS],
+            m_chunks_used: [0; (XSUM_LCHUNKS / 64)],
             m_used_used: 0,
             m_sacc: SmallAccumulator::new(),
         }
     }
 
-    pub(crate) fn add_lchunk_to_small(&mut self, ix: i32) {
-        let count: i32 = self.m_count[ix as usize];
+    pub(crate) fn add_lchunk_to_small(&mut self, ix: usize) {
+        let count: i32 = self.m_count[ix];
 
         // Add to the small accumulator only if the count is not -1, which
         // indicates a chunk that contains nothing yet.
@@ -39,7 +39,7 @@ impl LargeAccumulator {
             // Get the chunk we will add.  Note that this chunk is the integer sum
             // of entire 64-bit floating-point representations, with sign, exponent,
             // and mantissa, but we want only the sum of the mantissas.
-            let mut chunk = self.m_chunk[ix as usize];
+            let mut chunk = self.m_chunk[ix];
 
             // If we added the maximum number of values to 'chunk', the sum of
             // the sign and exponent parts (all the same, equal to the index) will
@@ -57,9 +57,9 @@ impl LargeAccumulator {
             // accumulator.  Noting that for denormalized numbers where the
             // exponent part is zero, the actual exponent is 1 (before subtracting
             // the bias), not zero.
-            let exp: i32 = ix & XSUM_EXP_MASK as i32;
+            let exp: i32 = ix as i32 & XSUM_EXP_MASK as i32;
             let mut low_exp: i32 = exp & XSUM_LOW_EXP_MASK as i32;
-            let mut high_exp: i32 = exp >> XSUM_LOW_EXP_BITS;
+            let mut high_exp: usize = (exp >> XSUM_LOW_EXP_BITS) as usize;
             if exp == 0 {
                 low_exp = 1;
                 high_exp = 0;
@@ -68,26 +68,26 @@ impl LargeAccumulator {
             // Split the mantissa into three parts, for three consecutive chunks in
             // the small accumulator.  Except for denormalized numbers, add in the sum
             // of all the implicit 1 bits that are above the actual mantissa bits.
-            let low_chunk: u64 = (chunk << low_exp) & XSUM_LOW_MANTISSA_MASK as u64;
-            let mut mid_chunk: u64 = chunk >> (XSUM_LOW_MANTISSA_BITS - low_exp as i64);
+            let low_chunk: i64 = (chunk << low_exp) as i64 & XSUM_LOW_MANTISSA_MASK;
+            let mut mid_chunk: i64 = chunk as i64 >> (XSUM_LOW_MANTISSA_BITS - low_exp as i64);
             if exp != 0 {
                 // normalized
-                mid_chunk += ((1 << XSUM_LCOUNT_BITS) as u64 - count as u64)
+                mid_chunk += ((1 << XSUM_LCOUNT_BITS) as i64 - count as i64)
                     << (XSUM_MANTISSA_BITS - XSUM_LOW_MANTISSA_BITS + low_exp as i64);
             }
-            let high_chunk: u64 = mid_chunk >> XSUM_LOW_MANTISSA_BITS;
-            mid_chunk &= XSUM_LOW_MANTISSA_MASK as u64;
+            let high_chunk: i64 = mid_chunk >> XSUM_LOW_MANTISSA_BITS;
+            mid_chunk &= XSUM_LOW_MANTISSA_MASK;
 
             // Add or subtract the three parts of the mantissa from three small
             // accumulator chunks, according to the sign that is part of the index.
             if ix & (1 << XSUM_EXP_BITS) != 0 {
-                self.m_sacc.m_chunk[high_exp as usize] -= low_chunk as i64;
-                self.m_sacc.m_chunk[(high_exp + 1) as usize] -= mid_chunk as i64;
-                self.m_sacc.m_chunk[(high_exp + 2) as usize] -= high_chunk as i64;
+                self.m_sacc.m_chunk[high_exp] -= low_chunk;
+                self.m_sacc.m_chunk[high_exp + 1] -= mid_chunk;
+                self.m_sacc.m_chunk[high_exp + 2] -= high_chunk;
             } else {
-                self.m_sacc.m_chunk[high_exp as usize] += low_chunk as i64;
-                self.m_sacc.m_chunk[(high_exp + 1) as usize] += mid_chunk as i64;
-                self.m_sacc.m_chunk[(high_exp + 2) as usize] += high_chunk as i64;
+                self.m_sacc.m_chunk[high_exp] += low_chunk;
+                self.m_sacc.m_chunk[high_exp + 1] += mid_chunk;
+                self.m_sacc.m_chunk[high_exp + 2] += high_chunk;
             }
 
             // The above additions/subtractions reduce by one the number we can
@@ -99,19 +99,19 @@ impl LargeAccumulator {
         // of adds we can do before the mantissa would overflow.  We also
         // set the bit in chunks_used to indicate that this chunk is in use
         // (if that is enabled).
-        self.m_chunk[ix as usize] = 0;
-        self.m_count[ix as usize] = 1 << XSUM_LCOUNT_BITS;
-        self.m_chunks_used[(ix >> 6) as usize] |= 1u64 << (ix & 0x3f);
+        self.m_chunk[ix] = 0;
+        self.m_count[ix] = 1 << XSUM_LCOUNT_BITS;
+        self.m_chunks_used[ix >> 6] |= 1u64 << (ix & 0x3f);
         self.m_used_used |= 1u64 << (ix >> 6);
     }
 
-    pub(crate) fn large_add_value_inf_nan(&mut self, ix: i32, uintv: u64) {
+    pub(crate) fn large_add_value_inf_nan(&mut self, ix: usize, uintv: u64) {
         if (ix as i64 & XSUM_EXP_MASK) == XSUM_EXP_MASK {
             self.m_sacc.add_inf_nan(uintv as i64);
         } else {
             self.add_lchunk_to_small(ix);
-            self.m_count[ix as usize] -= 1;
-            self.m_chunk[ix as usize] += uintv;
+            self.m_count[ix] -= 1;
+            self.m_chunk[ix] += uintv;
         }
     }
 
@@ -177,7 +177,7 @@ impl LargeAccumulator {
             // Find and process the chunks in this block that are used.  We skip
             // forward based on the m_chunksUsed flags until we're within eight
             // bits of a chunk that is in use.
-            let mut ix: i32 = (p << 6) as i32;
+            let mut ix: usize = p << 6;
             if (u & 0xffffffff) == 0 {
                 u >>= 32;
                 ix += 32;
@@ -192,7 +192,7 @@ impl LargeAccumulator {
             }
 
             loop {
-                if self.m_count[ix as usize] >= 0 {
+                if self.m_count[ix] >= 0 {
                     self.add_lchunk_to_small(ix);
                 }
                 ix += 1;
